refactor: move story providers to global decorator
